moved auction proposals from streamer to OTCbuyback
