+++ x/otcbuyback/keeper/msg_server.go
@@ -7,9 +7,12 @@ import (
 	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
 	"github.com/dymensionxyz/dymension/v3/x/otcbuyback/types"
+	streamertypes "github.com/dymensionxyz/dymension/v3/x/streamer/types"
+	"github.com/dymensionxyz/gerr-cosmos/gerrc"
 )
 
 type msgServer struct {
@@ -75,6 +78,67 @@ func (ms msgServer) SetAcceptedTokens(goCtx context.Context, req *types.MsgSetAc
 	}
 
 	return &types.MsgSetAcceptedTokensResponse{}, nil
+}
+
+// CreateAuction implements the MsgServer interface for creating Dutch auctions
+func (s msgServer) CreateAuction(goCtx context.Context, msg *types.MsgCreateAuction) (*types.MsgCreateAuctionResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	if msg.Authority != s.authority {
+		return nil, errorsmod.Wrapf(gerrc.ErrUnauthenticated, "invalid authority; expected %s, got %s", s.authority, msg.Authority)
+	}
+
+	k := s.Keeper
+
+	// validate streamer module has enough tokens to fund the auction
+	moduleBalance := k.bankKeeper.GetAllBalances(ctx, authtypes.NewModuleAddress(streamertypes.ModuleName))
+	alreadyAllocatedCoins := k.streamerKeeper.GetModuleToDistributeCoins(ctx)
+	if !sdk.NewCoins(msg.Allocation).IsAllLTE(moduleBalance.Sub(alreadyAllocatedCoins...)) {
+		return nil, fmt.Errorf("insufficient module balance to distribute coins")
+	}
+
+	// move funds from the streamer account to the auction account
+	err := k.bankKeeper.SendCoinsFromModuleToModule(ctx, streamertypes.ModuleName, types.ModuleName, sdk.NewCoins(msg.Allocation))
+	if err != nil {
+		return nil, err
+	}
+
+	// Create the auction and stream
+	auctionID, err := k.CreateAuction(
+		ctx,
+		msg.Allocation,
+		msg.StartTime,
+		msg.EndTime,
+		msg.InitialDiscount,
+		msg.MaxDiscount,
+		types.Auction_VestingParams{
+			VestingPeriod:               msg.VestingPeriod,
+			VestingStartAfterAuctionEnd: msg.VestingStartAfterAuctionEnd,
+		},
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return &types.MsgCreateAuctionResponse{
+		AuctionId: auctionID,
+	}, nil
+}
+
+// TerminateAuction implements the MsgServer interface for terminating Dutch auctions
+func (s msgServer) TerminateAuction(goCtx context.Context, msg *types.MsgTerminateAuction) (*types.MsgTerminateAuctionResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	if msg.Authority != s.authority {
+		return nil, errorsmod.Wrapf(gerrc.ErrUnauthenticated, "invalid authority; expected %s, got %s", s.authority, msg.Authority)
+	}
+
+	err := s.Keeper.EndAuction(ctx, msg.AuctionId, "auction_terminated")
+	if err != nil {
+		return nil, err
+	}
+
+	return &types.MsgTerminateAuctionResponse{}, nil
 }
 
 // Buy handles token purchase requests
