+++ x/otcbuyback/types/codec.go
@@ -10,6 +10,8 @@ import (
 // RegisterCodec registers the necessary x/otcbuyback interfaces and concrete types on the provided
 // LegacyAmino codec. These types are used for Amino JSON serialization.
 func RegisterCodec(cdc *codec.LegacyAmino) {
+	cdc.RegisterConcrete(&MsgCreateAuction{}, "otcbuyback/MsgCreateAuction", nil)
+	cdc.RegisterConcrete(&MsgTerminateAuction{}, "otcbuyback/MsgTerminateAuction", nil)
 	cdc.RegisterConcrete(&MsgBuy{}, "otcbuyback/MsgBuy", nil)
 	cdc.RegisterConcrete(&MsgBuyExactSpend{}, "otcbuyback/MsgBuyExactSpend", nil)
 	cdc.RegisterConcrete(&MsgClaimTokens{}, "otcbuyback/MsgClaimTokens", nil)
@@ -21,6 +23,8 @@ func RegisterCodec(cdc *codec.LegacyAmino) {
 func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
 	registry.RegisterImplementations(
 		(*sdk.Msg)(nil),
+		&MsgCreateAuction{},
+		&MsgTerminateAuction{},
 		&MsgBuy{},
 		&MsgBuyExactSpend{},
 		&MsgClaimTokens{},
